# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Asset Management System API (ISO 27001)
  version: 1.0.0
  description: RESTful API for IT Asset & ISMS Management with ISO 27001 compliance
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api/v1
    description: Development
paths:
...
  /controls/implementations:
    post:
      summary: Create scoped control implementation fulfilling multiple requirements
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ControlImplementationInput' }
      responses:
        '201': { description: Control implementation with findings and actions }
  /controls/soa:
    get:
      summary: List Statement of Applicability versions with SoAItems
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: SoA list }
    post:
      summary: Create Statement of Applicability from individual SoAItems
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SoAInput' }
      responses:
        '201': { description: Draft SoA }
  /controls/soa/{id}/submit:
    post:
      summary: Submit SoA for formal approval
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: SoA under review }
  /controls/soa/{id}/approve:
    post:
      summary: Approve or reject SoA; approved versions become immutable
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                decision: { type: string, enum: [approved, rejected], default: approved }
                comment: { type: string }
      responses:
        '200': { description: Approval decision recorded }
  /evidence:
    get:
      summary: List evidence metadata and relations
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Evidence list }
    post:
      summary: Create secure evidence metadata with SHA-256 hash and retention
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EvidenceInput' }
      responses:
        '201': { description: Evidence metadata }
  /evidence/audit-package:
    post:
      summary: Export evidence audit package manifest
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Audit package manifest }
  /evidence/{id}:
    delete:
      summary: Archive evidence unless delete-protected or retained
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Evidence archived }
        '409': { description: Evidence protected from deletion }
  /documents:
    post:
      summary: Create controlled policy document with draft version
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PolicyDocumentInput' }
      responses:
        '201': { description: Controlled document }
  /documents/{id}/transition:
    post:
      summary: Transition document workflow draft-review-approved-published-withdrawn
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Workflow transition recorded }
  /documents/{id}/acknowledge:
    post:
      summary: Acknowledge approved or published document
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Acknowledgement recorded }
  /documents/reviews/escalate-overdue:
    post:
      summary: Escalate overdue document reviews
      tags: [Phase4]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Escalation result }

  # ==================== Health ====================
...
  /auth/create-first-admin:
    post:
      summary: Create first admin user (one-time setup)
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/register:
    post:
      summary: Register new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already registered
  /auth/login:
    post:
      summary: Login with email/password
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      summary: Get current user profile
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /auth/me/preferences:
    patch:
      summary: Update user preferences (language, darkMode)
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language: { type: string, example: de }
                darkMode: { type: boolean }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /auth/refresh:
    post:
      summary: Refresh JWT token
      tags: [Authentication]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }

  # ==================== OIDC ====================
...
  /auth/oidc/callback:
    post:
      summary: OIDC callback handler
      tags: [OIDC]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, state]
              properties:
                code: { type: string }
                state: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/oidc/config:
    get:
      summary: Get OIDC configuration (safe, no secrets)
      tags: [OIDC]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OidcConfigResponse'

  # ==================== Admin - Users ====================
...
  /controls/tests:
    post:
      summary: Create a control implementation test with generic evidence links
      tags: [Controls]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ControlTestInput'
      responses:
        '201': { description: Control test created }
...
  /controls:
    get:
      summary: List controls
      tags: [Controls]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  # ==================== Incidents ====================
...
  /contracts:
    get:
      summary: List contracts
      tags: [Contracts]
      security: [{ bearerAuth: [] }]

  # ==================== Licenses ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ControlImplementationInput:
      type: object
      required: [controlId, responsibleUserId]
      properties:
        controlId: { type: string, format: uuid }
        scopeId: { type: string, format: uuid }
        organizationUnitId: { type: string, format: uuid }
        siteId: { type: string, format: uuid }
        responsibleUserId: { type: string }
        maturityLevel: { type: integer, minimum: 0, maximum: 5 }
        testMethod: { type: string }
        testFrequency: { type: string }
        nextTestDate: { type: string, format: date-time }
        requirementIds: { type: array, items: { type: string, format: uuid } }
...
    ControlTestInput:
      type: object
      required: [controlImplementationId, testType, testedBy, result]
      properties:
        controlImplementationId: { type: string, format: uuid }
        testType: { type: string }
        testMethod: { type: string }
        testedBy: { type: string }
        testedAt: { type: string, format: date-time }
        result: { type: string }
        effectivenessRating: { type: integer, minimum: 0, maximum: 100 }
        findings: { type: string }
        evidenceRequired: { type: boolean }
        nextTestDate: { type: string, format: date-time }
        evidenceLinks:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceLinkInput'
    EvidenceLinkInput:
      type: object
      required: [evidenceId]
      properties:
        evidenceId: { type: string, format: uuid }
        relationType: { type: string, default: supports }
    SoAInput:
      type: object
      required: [frameworkId, frameworkVersion, scopeId]
      properties:
        frameworkId: { type: string, format: uuid }
        frameworkVersion: { type: string }
        scopeId: { type: string }
        items:
          type: array
          items:
            type: object
            required: [justification]
            properties:
              requirementId: { type: string, format: uuid }
              controlId: { type: string, format: uuid }
              applicability: { type: string, enum: [applicable, not_applicable, under_review] }
              justification: { type: string }
              implementationStatus: { type: string }
              controlImplementationIds: { type: array, items: { type: string, format: uuid } }
    EvidenceInput:
      type: object
      required: [title, evidenceType, classification, responsibleId, fileHash]
      properties:
        title: { type: string }
        evidenceType: { type: string }
        classification: { type: string }
        fileHash: { type: string, minLength: 64, maxLength: 64 }
        links:
          type: array
          description: Generic evidence associations replacing relatedControlIds/relatedRiskIds mirror arrays.
          items:
            type: object
            required: [entityType, entityId]
            properties:
              entityType: { type: string, enum: [Control, Risk, Asset, SoAItem, Document, RiskControlAssessment, ControlTest] }
              entityId: { type: string }
              relationType: { type: string }
...
    PolicyDocumentInput:
      type: object
      required: [title, documentType, ownerId, content]
      properties:
        title: { type: string }
        documentType: { type: string }
        ownerId: { type: string }
        reviewerId: { type: string }
        approverId: { type: string }
        nextReviewDate: { type: string, format: date-time }
        reviewIntervalDays: { type: integer }
        content: { type: string }
        changeLog: { type: string }
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    RegisterInput:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12 }
        firstName: { type: string }
        lastName: { type: string }
        phoneNumber: { type: string }
        organizationUnitId: { type: string }
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        token: { type: string, description: JWT Bearer Token }
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayId: { type: string }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phoneNumber: { type: string, nullable: true }
        isActive: { type: boolean }
        roles:
          type: array
          items: { type: string }
        language: { type: string, nullable: true }
        darkMode: { type: boolean, nullable: true }
        createdAt: { type: string, format: date-time }
...
    OidcConfigResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        enabled: { type: boolean }
        providerName: { type: string }
        alloweEmailDomains:
          type: array
          items: { type: string }
        autoProvisioning: { type: boolean }
        defaultRoleForNewUsers: { type: string }
        enableLocalLogin: { type: boolean }
...
    PaginatedResponse:
      type: object
      properties:
        data: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }

# ==================== Phase 8: API, Betrieb und Nachweise ====================

  # ==================== Health & System (Phase 8) ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== Phase 8 Schemas ====================
...